refactor: juntando todos os notFoundExceptions em um único pacote
